Move Domain class under Spree::GraphQL namespace
